Move articles to docs and update imports
