+++ docs/articles/monthly-generator-pm-checklist.md
